Add --path CLI argument
Sets the path to the directory containing the images to be served.

+++ src/bin/laya-server.rs
@@ -1,14 +1,26 @@
-use std::path::PathBuf;
+use std::path::Path;
+
+use clap::Parser;
 
 #[path = "laya-server/telemetry.rs"]
 mod telemetry;
 
 fn main() {
+    let args = Args::parse();
+    if !args.path.exists() {
+        panic!("Could not resolve image directory '{}'.", args.path.to_string_lossy())
+    }
+
     if let Err(e) = telemetry::install_telemetry() {
         eprintln!("Failed to install telemetry {:?}", e);
     }
 
-    let mut path = PathBuf::new();
-    path.push("./share");
-    laya::start(path.into_boxed_path());
+    laya::start(args.path);
+}
+
+#[derive(Parser, Debug)]
+#[command(version, about, long_about = None)]
+struct Args {
+    #[arg(short, long, default_missing_value = "./share/")]
+    path: Box<Path>,
 }
